+++ diffsims/libraries/structure_library.py
@@ -81,36 +81,6 @@ class StructureLibrary:
         """
         return cls(identifiers, structures, orientations)
 
-    @classmethod
-    def from_crystal_systems(
-        cls, identifiers, structures, systems, resolution, equal="angle"
-    ):
-        """
-        Creates a structure library from crystal system derived orientation lists
-
-        Parameters
-        ----------
-        identifiers : list of strings/ints
-            A list of phase identifiers referring to different atomic structures.
-        structures : list of diffpy.structure.Structure objects.
-            A list of diffpy.structure.Structure objects describing the atomic
-            structure associated with each phase in the library.
-        systems : list
-            A list over indentifiers of crystal systems
-        resolution : float
-            resolution in degrees
-        equal : str
-            Default is 'angle'
-
-        Raises
-        ------
-        NotImplementedError:
-            "This function has been removed in version 0.3.0, in favour of creation from orientation lists"
-        """
-        raise NotImplementedError(
-            "This function has been removed in version 0.3.0, in favour of creation from orientation lists"
-        )
-
     def get_library_size(self, to_print=False):
         """
         Returns the the total number of orientations in the